refactor: notification 관련 로직 hook으로 분리
